update: replace selectedAccount portfolio and defi banners with balanceAffectingErrors
