+++ src/libs/banners/banners.ts
@@ -2,20 +2,9 @@ import { Account } from '../../interfaces/account'
 import { AccountOpAction, Action as ActionFromActionsQueue } from '../../interfaces/actions'
 // eslint-disable-next-line import/no-cycle
 import { Action, Banner } from '../../interfaces/banner'
-import { Network, NetworkId } from '../../interfaces/network'
-import { RPCProviders } from '../../interfaces/provider'
-import { SelectedAccountPortfolioState } from '../../interfaces/selectedAccount'
+import { Network } from '../../interfaces/network'
 import { ActiveRoute } from '../../interfaces/swapAndBridge'
-import {
-  AccountState as DefiPositionsAccountState,
-  DeFiPositionsError,
-  NetworksWithPositions
-} from '../defiPositions/types'
-import { getNetworksWithFailedRPC } from '../networks/networks'
-import { PORTFOLIO_LIB_ERROR_NAMES } from '../portfolio/portfolio'
 import { getIsBridgeTxn, getQuoteRouteSteps } from '../swapAndBridge/swapAndBridge'
-
-const TEN_MINUTES = 10 * 60 * 1000
 
 const getBridgeBannerTitle = (routeStatus: ActiveRoute['routeStatus']) => {
   switch (routeStatus) {
@@ -322,276 +311,4 @@ export const getKeySyncBanner = (addr: string, email: string, keys: string[]) =>
     ]
   }
   return banner
-}
-
-export const getNetworksWithFailedRPCBanners = ({
-  providers,
-  networks,
-  networksWithAssets
-}: {
-  providers: RPCProviders
-  networks: Network[]
-  networksWithAssets: NetworkId[]
-}): Banner[] => {
-  const banners: Banner[] = []
-  const networkIds = getNetworksWithFailedRPC({ providers }).filter((networkId) =>
-    networksWithAssets.includes(networkId)
-  )
-
-  const networksData = networkIds.map((id) => networks.find((n: Network) => n.id === id)!)
-
-  const allFailed = networksData.length === networks.length
-
-  const networksWithMultipleRpcUrls = allFailed
-    ? []
-    : networksData.filter((n) => n?.rpcUrls?.length > 1)
-
-  const networksToGroupInSingleBanner = allFailed
-    ? networksData
-    : networksData.filter((n) => n?.rpcUrls?.length <= 1)
-
-  if (!networksData.length) return banners
-
-  networksWithMultipleRpcUrls.forEach((n) => {
-    banners.push({
-      id: `${n.id}-custom-rpcs-down`,
-      type: 'error',
-      title: `Failed to retrieve network data for ${n.name}. You can try selecting another RPC URL`,
-      text: 'Affected features: visible assets, DeFi positions, sign message/transaction, ENS/UD domain resolving, add account.',
-      actions: [
-        {
-          label: 'Select',
-          actionName: 'select-rpc-url',
-          meta: {
-            network: n
-          }
-        }
-      ]
-    })
-  })
-
-  if (!networksToGroupInSingleBanner.length) return banners
-
-  banners.push({
-    id: 'rpcs-down',
-    type: 'error',
-    title: `Failed to retrieve network data for ${networksToGroupInSingleBanner
-      .map((n) => n.name)
-      .join(', ')} (RPC malfunction)`,
-    text: 'Affected features: visible assets, DeFi positions, sign message/transaction, ENS/UD domain resolving, add account. Please try again later or contact support.',
-    actions: []
-  })
-
-  return banners
-}
-
-export const getNetworksWithPortfolioErrorBanners = ({
-  networks,
-  selectedAccountLatest,
-  providers
-}: {
-  networks: Network[]
-  selectedAccountLatest: SelectedAccountPortfolioState
-  providers: RPCProviders
-}): Banner[] => {
-  const banners: Banner[] = []
-
-  const portfolioLoading = Object.keys(selectedAccountLatest).some((network) => {
-    const portfolioForNetwork = selectedAccountLatest[network]
-
-    return portfolioForNetwork?.isLoading
-  })
-
-  // Otherwise networks are appended to the banner one by one, which looks weird
-  if (portfolioLoading) return []
-
-  const networkNamesWithCriticalError: string[] = []
-  const networkNamesWithPriceFetchError: string[] = []
-
-  if (!Object.keys(selectedAccountLatest).length) return []
-
-  Object.keys(selectedAccountLatest).forEach((network) => {
-    const portfolioForNetwork = selectedAccountLatest[network]
-    const criticalError = portfolioForNetwork?.criticalError
-    const lastSuccessfulUpdate = portfolioForNetwork?.result?.lastSuccessfulUpdate
-
-    // Don't display an error banner if the last successful update was less than 10 minutes ago
-    if (typeof lastSuccessfulUpdate === 'number' && Date.now() - lastSuccessfulUpdate < TEN_MINUTES)
-      return
-
-    let networkName: string | null = null
-
-    if (network === 'gasTank') networkName = 'Gas Tank'
-    else if (network === 'rewards') networkName = 'Rewards'
-    else networkName = networks.find((n) => n.id === network)?.name ?? null
-
-    if (!portfolioForNetwork || !networkName || portfolioForNetwork.isLoading) return
-    // Don't display an error banner if the RPC isn't working because an RPC error banner is already displayed.
-    // In case of additional networks don't check the RPC as there isn't one
-    if (
-      criticalError &&
-      (['gasTank', 'rewards'].includes(network) || providers[network].isWorking)
-    ) {
-      networkNamesWithCriticalError.push(networkName as string)
-      // If there is a critical error, we don't need to check for price fetch error
-      return
-    }
-
-    portfolioForNetwork?.errors.forEach((err: any) => {
-      // If the error is not predefined(shouldn't happen) or it's a specific
-      // non-critical error, we shouldn't display a banner
-      if (
-        !Object.keys(PORTFOLIO_LIB_ERROR_NAMES).includes(err?.name) ||
-        err?.name === PORTFOLIO_LIB_ERROR_NAMES.NonCriticalApiHintsError
-      )
-        return
-
-      // Price errors are not critical, so we should display a warning banner
-      if (err?.name === PORTFOLIO_LIB_ERROR_NAMES.PriceFetchError) {
-        networkNamesWithPriceFetchError.push(networkName as string)
-        return
-      }
-
-      // For all other errors, we should display an error banner
-      networkNamesWithCriticalError.push(networkName as string)
-    })
-  })
-
-  if (networkNamesWithPriceFetchError.length) {
-    banners.push({
-      id: 'portfolio-prices-error',
-      type: 'warning',
-      title: `Failed to retrieve prices for ${networkNamesWithPriceFetchError.join(', ')}`,
-      text: 'Affected features: account balances, asset prices. Reload the account or try again later.',
-      actions: []
-    })
-  }
-  if (networkNamesWithCriticalError.length) {
-    banners.push({
-      id: 'portfolio-critical-error',
-      type: 'error',
-      title: `Failed to retrieve the portfolio data for ${networkNamesWithCriticalError.join(
-        ', '
-      )}`,
-      text: 'Affected features: account balances, visible assets. Reload the account or try again later.',
-      actions: []
-    })
-  }
-
-  return banners
-}
-
-export const getNetworksWithDeFiPositionsErrorBanners = ({
-  networks,
-  currentAccountState,
-  providers,
-  networksWithPositions
-}: {
-  networks: Network[]
-  currentAccountState: DefiPositionsAccountState
-  providers: RPCProviders
-  networksWithPositions: NetworksWithPositions
-}) => {
-  const isLoading = Object.keys(currentAccountState).some((networkId) => {
-    const networkState = currentAccountState[networkId]
-    return networkState.isLoading
-  })
-
-  if (isLoading) return []
-
-  const networkNamesWithUnknownCriticalError: string[] = []
-  const networkNamesWithAssetPriceCriticalError: string[] = []
-  const providersWithErrors: {
-    [providerName: string]: string[]
-  } = {}
-
-  Object.keys(currentAccountState).forEach((networkId) => {
-    const providersWithPositions = networksWithPositions[networkId]
-    // Ignore networks that don't have positions
-    // but ensure that we have a successful response stored (the network key is present)
-    if (providersWithPositions && !providersWithPositions.length) return
-
-    const networkState = currentAccountState[networkId]
-    const network = networks.find((n) => n.id === networkId)
-    const rpcProvider = providers[networkId]
-    const lastSuccessfulUpdate = networkState.updatedAt
-
-    if (
-      !network ||
-      !networkState ||
-      (typeof lastSuccessfulUpdate === 'number' &&
-        Date.now() - lastSuccessfulUpdate < TEN_MINUTES) ||
-      // Don't display an error banner if the RPC isn't working because an RPC error banner is already displayed.
-      (typeof rpcProvider.isWorking === 'boolean' && !rpcProvider.isWorking)
-    )
-      return
-
-    if (networkState.error) {
-      if (networkState.error === DeFiPositionsError.AssetPriceError) {
-        networkNamesWithAssetPriceCriticalError.push(network.name)
-      } else if (networkState.error === DeFiPositionsError.CriticalError) {
-        networkNamesWithUnknownCriticalError.push(network.name)
-      }
-    }
-
-    const providerNamesWithErrors =
-      networkState.providerErrors
-        ?.filter(({ providerName }) => {
-          // Display all errors if there hasn't been a successful update
-          // for the network.
-          if (!networksWithPositions[networkId]) return true
-          // Exclude providers without positions
-          return networksWithPositions[networkId].includes(providerName)
-        })
-        .map((e) => e.providerName) || []
-
-    if (providerNamesWithErrors.length) {
-      providerNamesWithErrors.forEach((providerName) => {
-        if (!providersWithErrors[providerName]) providersWithErrors[providerName] = []
-
-        providersWithErrors[providerName].push(network.name)
-      })
-    }
-  })
-
-  const providerErrorBanners: Banner[] = Object.entries(providersWithErrors).map(
-    ([providerName, networkNames]) => {
-      return {
-        id: `${providerName}-defi-positions-error`,
-        type: 'error',
-        title: `Failed to retrieve DeFi positions for ${providerName} on ${networkNames.join(
-          ', '
-        )}`,
-        text: 'Reload the account or try again later.',
-        actions: []
-      }
-    }
-  )
-
-  const banners = providerErrorBanners
-
-  if (networkNamesWithUnknownCriticalError.length) {
-    banners.push({
-      id: 'defi-positions-critical-error',
-      type: 'error',
-      title: `Failed to retrieve DeFi positions on ${networkNamesWithUnknownCriticalError.join(
-        ', '
-      )}`,
-      text: 'Reload the account or try again later.',
-      actions: []
-    })
-  }
-  if (networkNamesWithAssetPriceCriticalError.length) {
-    banners.push({
-      id: 'defi-positions-asset-price-error',
-      type: 'warning',
-      title: `Failed to retrieve asset prices for DeFi positions on ${networkNamesWithAssetPriceCriticalError.join(
-        ', '
-      )}`,
-      text: 'Reload the account or try again later.',
-      actions: []
-    })
-  }
-
-  return banners
 }
